feat: add macOS service installer and rename project to remindctl-mcp
Make it easier to run the MCP server persistently on macOS by adding launchd install/uninstall scripts and clear service setup docs. Align package naming and runtime behavior with deployment expectations so remindctl runs reliably under launchd.

+++ src/models.rs
@@ -84,8 +84,8 @@ pub struct BatchActionResult {
 
 #[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
 pub struct BatchProcessResult {
-    pub processed: usize,
-    pub succeeded: usize,
-    pub failed: usize,
+    pub processed: i64,
+    pub succeeded: i64,
+    pub failed: i64,
     pub results: Vec<BatchActionResult>,
 }
